+++ base32_from.c
@@ -96,6 +96,6 @@ int base32_valid(const char *src,size_t *count)
 		;
 
 	if (count)
-		*count = p - src;
+		*count = (size_t) (p - src);
 	return !*p;
 }
